♻️ Group backend adapters by technology (Task_4)
infrastructures/ dissolves into adapters/{oxigraph,qdrant,openai,stats}
under one grouping scheme. The Oxigraph adapter file splits into
embedded.rs, http.rs, and shared.rs with its test module alongside.
All three RetrievalStatsStore implementations (noop, in-memory,
sqlite) co-locate under adapters/stats. Vector value types move to
src/models, embedding surfaces to src/policy.
