+++ src/adapters.rs
@@ -1,0 +1,14 @@
+#![allow(unused_imports)]
+
+pub(crate) mod openai;
+pub(crate) mod oxigraph;
+pub(crate) mod qdrant;
+pub(crate) mod stats;
+
+pub(crate) use openai::OpenAIEmbeddingProvider;
+pub(crate) use oxigraph::{OxigraphGraphAuthorityStore, OxigraphHttpGraphAuthorityStore};
+pub(crate) use qdrant::QdrantVectorCandidateStore;
+pub(crate) use stats::{InMemoryRetrievalStatsStore, SqliteRetrievalStatsStore};
+
+#[cfg(test)]
+pub(crate) use stats::noop_retrieval_stats_store;
