+++ src/adapters/openai.rs
@@ -1,0 +1,3 @@
+mod embedding_provider;
+
+pub(crate) use embedding_provider::OpenAIEmbeddingProvider;
