+++ src/adapters/oxigraph.rs
@@ -1,0 +1,11 @@
+mod embedded;
+mod http;
+mod rdf_mapping;
+mod shared;
+mod sparql_selectors;
+#[cfg(test)]
+mod tests;
+mod vocabulary;
+
+pub(crate) use embedded::OxigraphGraphAuthorityStore;
+pub(crate) use http::OxigraphHttpGraphAuthorityStore;
