+++ src/adapters/oxigraph/vocabulary.rs
