+++ src/adapters/qdrant.rs
@@ -1,0 +1,4 @@
+mod payload;
+mod store;
+
+pub(crate) use store::QdrantVectorCandidateStore;
