+++ src/adapters/stats/sqlite.rs
