+++ src/internal.rs
@@ -1,4 +1,2 @@
 pub(crate) mod config;
-pub(crate) mod infrastructures;
-pub(crate) mod models;
 pub(crate) mod repositories;
