--- src/internal/infrastructures.rs
@@ -1,3 +1,0 @@
-pub(crate) mod external_services;
-pub(crate) mod graph;
-pub(crate) mod retrieval_stats;
