--- src/internal/infrastructures/external_services.rs
@@ -1,6 +1,0 @@
-mod openai_embedding_provider;
-mod qdrant_payload;
-mod qdrant_vector_candidate_store;
-
-pub(crate) use openai_embedding_provider::OpenAIEmbeddingProvider;
-pub(crate) use qdrant_vector_candidate_store::QdrantVectorCandidateStore;
