--- src/internal/infrastructures/graph.rs
@@ -1,8 +1,0 @@
-mod oxigraph_authority_store;
-mod rdf_mapping;
-mod sparql_selectors;
-mod vocabulary;
-
-pub(crate) use oxigraph_authority_store::{
-    OxigraphGraphAuthorityStore, OxigraphHttpGraphAuthorityStore,
-};
