+++ src/models/vector.rs
@@ -1,6 +1,5 @@
 mod candidate_record;
 mod embedding_model;
-mod embedding_surface;
 mod record;
 
 // Provider-neutral vector model surface. Adapters, pipelines, and test
@@ -12,11 +11,6 @@ pub(crate) use candidate_record::{
     VectorTimeField, VectorTimeRangeFilter,
 };
 pub(crate) use embedding_model::EmbeddingModel;
-#[allow(unused_imports)]
-pub(crate) use embedding_surface::{
-    derived_memory_vector_record, entity_vector_record, episode_vector_record,
-    memory_object_vector_record, memory_thread_vector_record, observation_vector_record,
-};
 #[allow(unused_imports)]
 pub(crate) use record::{
     VectorCandidateDiagnosticRecord, VectorPayloadHints, VectorRecord, VectorRecordEmbedding,
