+++ src/models/vector/candidate_record.rs
