+++ src/models/vector/embedding_model.rs
